Refactor code to allow common listing render functions, initial not-pretty rendering of search results with new model

+++ apps/shared/lib/model/listing/meta/standard/cost/minimum_purchase.ex
@@ -4,7 +4,7 @@ defmodule Standard.Cost.MinimumPurchase do
   @derive {Poison.Encoder, except: [:__meta__]}
   @primary_key false
   embedded_schema do
-    field :data, :integer
+    field :data, :float
     field :type, :string
   end
